resample audio (#53)
With samplerates of 44100 or other values that are not a multiple of 16000, amount of data sent for the time period is incorrect. Though not noticeable from audio file, if using AEC audio with sample rate of 48000, will not evenly match when sent. This results in an ever-growing buffer of audio data that will cause a slowly growing delay in audio processing